+++ atcg_lib/include/Renderer/RenderPasses/BlitPass.h
@@ -3,16 +3,16 @@
 #include <Renderer/RenderPass.h>
 #include <Renderer/Texture.h>
 #include <DataStructure/Skybox.h>
+#include <Scene/ComponentRenderer.h>
 
 namespace atcg
 {
+
 /**
- * @brief A RenderPass that renders a skybox.
+ * @brief A RenderPass that blits two framebuffer.
  *
  * This pass reads the following variables from the context:
  * * context<RendererSystem*>["renderer"] - The renderer
- * * context<bool>["has_skybox"] - If the skybox is present
- * * context<atcg::ref_ptr<Skybox>>["skybox"] - The skybox used
  *
  * data:
  * * data<atcg::ref_ptr<ref_ptr<Framebuffer>>>["target"] - The target if RenderTargetMode is set to
@@ -23,24 +23,18 @@ namespace atcg
  * RENDER_TARGET_INPUTFRAMEBUFFER
  *
  * outputs:
- * * outputs<ref_ptr<ref_ptr<Skybox>>>["skybox"] - The skybox used. Might be a dummy skybox
  * * outputs<ref_ptr<ref_ptr<Framebuffer>>>["framebuffer"] - The target framebuffer
  *
  */
-class SkyboxPass : public RenderPass
+class BlitPass : public RenderPass
 {
 public:
     /**
-     * @brief Constructor
-     */
-    SkyboxPass();
-
-    /**
-     * @brief Constructor
+     * @brief Constructor.
      *
-     * @param desc The render target description
+     * @param desc The Render target description
      */
-    SkyboxPass(const RenderTargetDesc& desc);
+    BlitPass(const RenderTargetDesc& desc = {});
 
 private:
     void initRenderPass();